requirements.txt
diffusers==0.21.4
einops==0.3.0
numpy==1.24.4
Pillow==9.4.0
pytorch_lightning==1.4.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
